moving files back to cmd/influxd package for now
